requirements.txt
aerich==0.7.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
charset-normalizer==3.1.0
click==8.1.3
dictdiffer==0.9.0
Django==4.2.2
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
multidict==6.0.4
psycopg==3.0.12
psycopg-binary==3.0.12
psycopg-pool==3.1.7
psycopg2-binary==2.9.6
py-cord==2.4.1
pydantic==1.10.9
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomlkit==0.11.8
tortoise-orm==0.19.3
typing_extensions==4.6.3
yarl==1.9.2
